requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2 ## to load any pdf 
chromadb ## vector storing 

